Rework how Time zones work
It appears setting the time zone via the bind was only setting it for the container shell and not the actual headless client, thus it really had no affect.
I have removed this, and instead added documentation for the inbuilt TZ variable. This is not a breaking change as when the TZ variable is missing like it is right now, it just defaults to UTC, which is what the headless is doing right now. However when it is set, the headless picks up the correct time zone which can be seen in the logs.